refactor(peagen): relocate internal modules into _utils
